Moved ProviderType to /common instead of /media3p
