+++ assets/src/edit-story/components/library/panes/media/common/providerType.js
