+++ assets/src/edit-story/components/library/panes/media/local/mediaPane.js
@@ -54,7 +54,7 @@ import {
   StyledPane,
 } from '../common/styles';
 import MediaGallery from '../common/mediaGallery';
-import { ProviderType } from '../media3p/providerType';
+import { ProviderType } from '../common/providerType';
 import paneId from './paneId';
 
 export const ROOT_MARGIN = 300;
